[ADD] integration dda + parser
