+++ map/valid/lodev.cub
@@ -1,0 +1,32 @@
+NO ./textures/onepiece06_chopper.xpm
+SO ./textures/building_takoyaki_yatai.xpm
+WE ./textures/building_kokuzeikyoku.xpm
+EA ./textures/animal_chara_radio_azarashi.xpm
+
+F 85,0,0
+C 135,206,235
+
+111111111111111111111111
+100000000000000000000001
+100000000000000000000001
+100000000000000000000001
+100000111110000101010001
+100000100010000000000001
+100000100010000100010001
+100000100010000000000001
+100000110110000101010001
+100000000000000000000001
+100000000000000000000001
+100000000000000000000001
+100000000000000000000001
+100000000000000000000001
+10000000000N000000000001
+100000000000000000000001
+111111111000000000000001
+110100001000000000000001
+110000101000000000000001
+110100001000000000000001
+110111111000000000000001
+110000000000000000000001
+111111111000000000000001
+111111111111111111111111
